requirements.txt
python-binance==1.0.29
tradingview-ta==3.3.0
PTable==0.9.2
colorama==0.4.6
PyYAML==6.0.2
requests==2.32.4
prettytable==3.16.0
scikit-learn==1.6.1
SQLAlchemy==2.0.42
loguru==0.7.3
pandas==2.3.1
numpy==2.3.2
dash==3.1.1
dash-ag-grid==32.3.0
dash-bootstrap-components==2.0.3
dash-auth==2.3.0
dash-iconify==0.1.2
docker==7.1.0
pytelegrambotapi==4.28.0
